Refactor: move dao related code to MoviesLoader
